# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tunploy API",
    "version": "1.2",
    "description": "Manage devices and servers on your Tunploy VPN from your own programs, and hear about what happens to them.\n\nSend an API key from Settings → API keys as `Authorization: Bearer tp_…`. Each key only does what its scopes allow. Requests are limited to 10 a second per key, with bursts up to 60. Keep keys on your server, never in an app.\n\nErrors always look like `{\"error\": {\"code\": \"…\", \"message\": \"…\", \"fields\": {…}}}`; `fields` names the inputs a validation error is about."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
    "/devices": {
      "get": {
        "operationId": "listDevices",
        "summary": "List devices",
        "tags": [
          "Devices"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:read",
        "responses": {
          "200": {
            "description": "Oldest first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevicePage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Needs the `devices:read` scope.",
        "parameters": [
          {
            "name": "server_id",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Only this server's devices."
          },
          {
            "name": "external_id",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Only devices with this external_id."
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DeviceStatus"
            },
            "description": "Only devices in this state."
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Cursor: the last ID of the previous page."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "description": "Page size."
          }
        ]
      },
      "post": {
        "operationId": "createDevice",
        "summary": "Create a device",
        "tags": [
          "Devices"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:write",
        "responses": {
          "201": {
            "description": "The device, with its config.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/Invalid"
          },
          "409": {
            "description": "`server_full`: no free addresses left; `no_server_available`: `server_id: \"auto\"` found no running server with room; or an Idempotency-Key is still in use.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Send `public_key` to keep the private key on the client; the config then has no PrivateKey line. Send `server_id: \"auto\"` (with `country` or `city` if you like) to let the panel pick the emptiest running server.\n\nNeeds the `devices:write` scope.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "A retry with the same key within 24 hours gets the first reply back (with `Idempotent-Replayed: true`) instead of running again."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "server_id"
                ],
                "properties": {
                  "server_id": {
                    "oneOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "const": "auto"
                      }
                    ],
                    "description": "A server's ID, or `\"auto\"` for the running server with the fewest devices that still has room, narrowed by `country` and `city`."
                  },
                  "country": {
                    "type": "string",
                    "description": "Only with `server_id: \"auto\"`: an ISO 3166-1 alpha-2 code such as `DE`."
                  },
                  "city": {
                    "type": "string",
                    "description": "Only with `server_id: \"auto\"`."
                  },
                  "public_key": {
                    "type": "string",
                    "description": "A WireGuard public key the client made."
                  },
                  "name": {
                    "type": "string",
                    "maxLength": 64,
                    "description": "Unique per server. Made up from external_id when left out."
                  },
                  "external_id": {
                    "type": "string",
                    "maxLength": 255
                  },
                  "metadata": {
                    "type": [
                      "object",
                      "null"
                    ],
                    "additionalProperties": true,
                    "description": "Any JSON object up to 4 KB; null clears it."
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "data_limit": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "limit_period": {
                    "$ref": "#/components/schemas/LimitPeriod"
                  },
                  "expires_at": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "date-time",
                    "description": "null clears it."
                  },
                  "speed_limit": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 10000000,
                    "description": "kbit/s (1000 bits a second), download and upload each; 0 means no limit."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/devices/{id}": {
      "get": {
        "operationId": "getDevice",
        "summary": "Get a device",
        "tags": [
          "Devices"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:read",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Needs the `devices:read` scope.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Device ID."
          }
        ]
      },
      "patch": {
        "operationId": "updateDevice",
        "summary": "Change a device",
        "tags": [
          "Devices"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:write",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/Invalid"
          },
          "502": {
            "description": "`apply_failed`: saved, but the running tunnel could not be updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Only the fields sent change. To put the device on another server use `POST /devices/{id}/move`; the public key cannot change.\n\nNeeds the `devices:write` scope.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Device ID."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 64,
                    "description": "Unique per server. Made up from external_id when left out."
                  },
                  "external_id": {
                    "type": "string",
                    "maxLength": 255
                  },
                  "metadata": {
                    "type": [
                      "object",
                      "null"
                    ],
                    "additionalProperties": true,
                    "description": "Any JSON object up to 4 KB; null clears it."
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "data_limit": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "limit_period": {
                    "$ref": "#/components/schemas/LimitPeriod"
                  },
                  "expires_at": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "date-time",
                    "description": "null clears it."
                  },
                  "speed_limit": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 10000000,
                    "description": "kbit/s (1000 bits a second), download and upload each; 0 means no limit."
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteDevice",
        "summary": "Delete a device",
        "tags": [
          "Devices"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:write",
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Needs the `devices:write` scope.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Device ID."
          }
        ]
      }
    },
    "/devices/{id}/config": {
      "get": {
        "operationId": "getDeviceConfig",
        "summary": "Get a device's config",
        "tags": [
          "Devices"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:read",
        "responses": {
          "200": {
            "description": "The config file, or a QR code of it.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "contentMediaType": "image/png"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "`client_key`: no QR code for a device that holds its own key; `split_tunnel`: a kill switch needs a server whose clients send all traffic through it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Needs the `devices:read` scope.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Device ID."
          },
          {
            "name": "format",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "conf",
                "qr"
              ],
              "default": "conf"
            },
            "description": ""
          },
          {
            "name": "kill_switch",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "`true` adds wg-quick firewall rules that block all traffic outside the tunnel while it is up, for Linux. Only with `format=conf`: phone and Windows apps refuse the PostUp lines, and have their own kill switch setting."
          }
        ]
      }
    },
    "/devices/{id}/usage": {
      "get": {
        "operationId": "getDeviceUsage",
        "summary": "Get a device's usage",
        "tags": [
          "Devices"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:read",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usage"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Needs the `devices:read` scope.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Device ID."
          }
        ]
      }
    },
    "/devices/{id}/usage/reset": {
      "post": {
        "operationId": "resetDeviceUsage",
        "summary": "Reset a device's usage",
        "tags": [
          "Devices"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:write",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "description": "`apply_failed`: reset, but the running tunnel could not be updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Starts the count toward the data limit again from now; a device blocked by its limit may connect at once. For plans that renew on their own date: use `limit_period: total` and reset on each renewal. History stays in the usage figures.\n\nNeeds the `devices:write` scope.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Device ID."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "A retry with the same key within 24 hours gets the first reply back (with `Idempotent-Replayed: true`) instead of running again."
          }
        ]
      }
    },
    "/devices/{id}/move": {
      "post": {
        "operationId": "moveDevice",
        "summary": "Move a device to another server",
        "tags": [
          "Devices"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:write",
        "responses": {
          "200": {
            "description": "The device, with its new config.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/Invalid"
          },
          "409": {
            "description": "`no_server_available`: `\"auto\"` found no running server with free addresses there; `server_full`: the chosen server has none left.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "`apply_failed`: moved, but a running tunnel could not be updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "The device keeps its ID, keys, name, limits and usage history. It gets an address on the new server, and the server's endpoint and key change, so the client needs the new `config` from the reply. With `server_id: \"auto\"` the server it is on now is never picked.\n\nNeeds the `devices:write` scope.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Device ID."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "A retry with the same key within 24 hours gets the first reply back (with `Idempotent-Replayed: true`) instead of running again."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "server_id"
                ],
                "properties": {
                  "server_id": {
                    "oneOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "const": "auto"
                      }
                    ],
                    "description": "A server's ID, or `\"auto\"` for the running server with the fewest devices that still has room, narrowed by `country` and `city`."
                  },
                  "country": {
                    "type": "string",
                    "description": "Only with `server_id: \"auto\"`: an ISO 3166-1 alpha-2 code such as `DE`."
                  },
                  "city": {
                    "type": "string",
                    "description": "Only with `server_id: \"auto\"`."
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/groups/{external_id}/usage/reset": {
      "post": {
        "operationId": "resetGroupUsage",
        "summary": "Reset usage for a group",
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:write",
        "responses": {
          "200": {
            "description": "The group after the reset.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "description": "`apply_failed`: reset, but a running tunnel could not be updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Resets every device in the group, as `POST /devices/{id}/usage/reset` does for one.\n\nNeeds the `devices:write` scope.",
        "parameters": [
          {
            "name": "external_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "The external_id the devices share, URL-encoded."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "A retry with the same key within 24 hours gets the first reply back (with `Idempotent-Replayed: true`) instead of running again."
          }
        ]
      }
    },
...
    "/nodes": {
      "get": {
        "operationId": "listNodes",
        "summary": "List nodes",
        "tags": [
          "Servers"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "servers:read",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodePage"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "The machines servers can run on. The panel's own is first, with ID 0. Nodes are added in the panel, since that takes SSH access.\n\nNeeds the `servers:read` scope."
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key: `tp_…`."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "`invalid_request`: a query parameter or the body could not be read.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "error": {
                "value": {
                  "error": {
                    "code": "invalid_request",
                    "message": "…"
                  }
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "`unauthorized`: no key, or it is unknown, revoked or expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "error": {
                "value": {
                  "error": {
                    "code": "unauthorized",
                    "message": "…"
                  }
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "`missing_scope`: the key lacks the scope this needs.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "error": {
                "value": {
                  "error": {
                    "code": "missing_scope",
                    "message": "…"
                  }
                }
              }
            }
          }
        }
      },
...
      "RateLimited": {
        "description": "`rate_limited`: wait for Retry-After seconds.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "error": {
                "value": {
                  "error": {
                    "code": "rate_limited",
                    "message": "…"
                  }
                }
              }
            }
          }
        },
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          }
        }
      }
    },
    "schemas": {
...
      "DeviceStatus": {
        "type": "string",
        "enum": [
          "active",
          "disabled",
          "expired",
          "limit_reached"
        ]
      },
...
      "DevicePage": {
        "type": "object",
        "required": [
          "data",
          "has_more"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Device"
            }
          },
          "has_more": {
            "type": "boolean",
            "description": "More items follow; ask again with the last ID as the cursor."
          }
        }
      },
...
    }
  }
}